move `InspectorProxyController` to its own file
